Adding the plisLib module to read xml plist representations into javascript objects and to serialize javascript objects into xml plist files.
Javascript knows only Number as numeric data type what is more float than int. So here is IntObject included to mark a number to be serialized as plist integer object. There's also a DataObject (to transport base64 encoded data) which is just a stub right now, because I have no testcase to do it right.