Fix test for faulty plist (fixes #13 and #14)
In Test_GlyphSet_rebuildContents_faultyPlist, check that IONoEntry is
thrown when trying to read a non-existent plist.

Remove Test_GlyphSet_rebuildContents_notFoundPlist, as its assumption,
that an empty glyphset will be returned for a non-existent plist, is now
false: an error is raised. This behaviour is in any case tested by
Test_GlyphSet_rebuildContents_faultyPlist.